restructured device group operations regarding messages
